Fix background-images on home page
