+++ src/assets/styles/home.scss
@@ -408,5 +408,3 @@
     }
   }
 }
-
-
