--- src/views/Home.vue
@@ -1,306 +1,0 @@
-<template>
-  <div>
-    <div class="agency-hero">
-      <section class="container">
-        <div class="row">
-          <div class="hero-text">
-            <p class="customFadeInUp">
-              We believe in simple and efficient ways to grow a business by being smart, relevant and most importantly straightforward.
-            </p>
-          </div>
-        </div>
-      </section>
-    </div>
-
-    <div class="index-slider-features">
-      <div class="container">
-        <div class="row">
-          <div class="col-md-3">
-            <div class="feature">
-              <h3>
-                <img src="~theme-corporate/images/icons/outline-color-bag.png" style="top:5px;" />
-                Great design for everyone
-              </h3>
-              <p>
-                Bring all your ideas together and make customizations faster and better than ever.
-              </p>
-              <a href="#">
-                Learn more
-              </a>
-            </div>
-          </div>
-          <div class="col-md-3">
-            <div class="feature">
-              <h3>
-                <img src="~theme-corporate/images/icons/outline-color-support.png" />
-
-                Offer support to your customers
-              </h3>
-              <p>
-                You can share your knowledge with your customers and everybody find their answers.
-              </p>
-              <a href="#">
-                Learn more
-              </a>
-            </div>
-          </div>
-          <div class="col-md-3">
-            <div class="feature">
-              <h3>
-                <img src="~theme-corporate/images/icons/outline-color-random.png" style="top:6px;" />
-
-                Quality code for customization
-              </h3>
-              <p>
-                Easily customize any part of your project for best results and quick great prototyping.
-              </p>
-              <a href="#">
-                Learn more
-              </a>
-            </div>
-          </div>
-          <div class="col-md-3">
-            <div class="feature">
-              <h3>
-                <img src="~theme-corporate/images/icons/outline-color-pen.png" style="top:6px;" />
-
-                Full responsiveness for any device
-              </h3>
-              <p>
-                Showcase your work on any computer or mobile device and all screen sizes.
-              </p>
-              <a href="#">
-                Learn more
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-  <div class="agency-portfolio">
-    <div class="container">
-      <h2>
-        Our latest projects
-      </h2>
-
-      <div class="row">
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>Landscape Trip</h3>
-                <p>
-                Blog About spacial trips
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1478059425650-ca13d6d422f4.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>App Icon</h3>
-                <p>
-                New icon for mobile app
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/project/shot-3.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>App Icon</h3>
-                <p>
-                New mobile app icon for fighters
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1467659226669-a1360d97be2d.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>Mobile Analytics</h3>
-                <p>
-                iOS mobile app for charts and stuff
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1474843148229-3163319fcc4c.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>App Food Icon</h3>
-                <p>
-                New icon for food mobile app
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1470274038469-958113db2384.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>App Food Icon</h3>
-                <p>
-                New icon for food mobile app
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/project/shot-2.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>Health App</h3>
-                <p>
-                Keep track of your health
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/project/shot-1.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>PSD Freebie</h3>
-                <p>
-                Free presentation mockups
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1467703834117-04386e3dadd8.jpg')});">
-            </span>
-          </a>
-        </div>
-        <div class="col-lg-4 col-md-6">
-          <a href="#" class="project">
-            <span class="mask">
-              <span class="info">
-                <h3>Grid Blog</h3>
-                <p>
-                Portfolio grid blog
-                </p>
-              </span>
-              <span class="btn-see-project">
-                See project
-              </span>
-            </span>
-            <span class="pic" style="background-image:url(${require('../../node_modules/theme-corporate/images/unsplash/photo-1479243396208-be1751535cdb.jpg')});">
-            </span>
-          </a>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="index-devices mobile-alt">
-    <section class="clearfix">
-      <img src="~theme-corporate/images/devices/devices-1.png" class="devices" />
-
-      <h4>Looks great in any device</h4>
-      <p>
-        Core has both a web app and an android app to make your website easy and always available. It offers you all the designs in collaboration with some smart people. Your projects will look great everywhere you go. Use new components that come included!
-      </p>
-      <p>
-        Core is an advanced theme solution for desktop, tablet, and mobile devices. It works in all browsers and your clients can pay instantly, to improve their experience at all times anywhere they go.
-      </p>
-      <div class="icons">
-        <i class="fa fa-desktop fa-3x"></i>
-        <i class="fa fa-tablet fa-3x"></i>
-        <i class="fa fa-mobile-phone fa-3x"></i>
-      </div>
-    </section>
-  </div>
-
-  <div class="spacial-client-quotes">
-    <section class="container">
-      <h2>
-        Used by the world's largest companies
-      </h2>
-      <Slidetext/>
-    </section>
-  </div>
-
-  <div class="call-to-action-signup">
-      <div class="container">
-        <h3>
-          Try it free for 30 days
-        </h3>
-        <form class="form-inline justify-content-center">
-          <div class="form-group">
-            <input type="text" class="form-control" placeholder="Your name" />
-          </div>
-          <div class="form-group">
-            <input type="email" class="form-control" placeholder="Your email address" />
-          </div>
-          <div class="form-group">
-            <input type="password" class="form-control" placeholder="Your password" />
-          </div>
-
-          <button class="btn-shadow btn-shadow-primary" type="submit">
-            Create account
-          </button>
-
-          <div class="agree w-100">By signing up you agree to our <a href="#">Terms</a>.</div>
-        </form>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script charset="utf-8">
-  import Slidetext from '@/components/Slidetext'
-
-  export default {
-    components: { Slidetext }
-  }
-</script>
