+++ src/views/home.html
@@ -242,6 +242,7 @@
   </div>
 </div>
 
+<<<<<<< HEAD
 <div class="index-devices mobile-alt">
   <section class="clearfix">
     <img src="~theme-corporate/images/devices/devices-1.png" class="devices" />
@@ -313,6 +314,123 @@
           <span class="role">Marketing, Atlassian</span>
         </div>
       </div>
+=======
+  <div class="index-devices mobile-alt">
+    <section class="clearfix">
+      <img src="~theme-corporate/images/devices/devices-1.png" class="devices" />
+
+      <h4>Looks great in any device</h4>
+      <p>
+        Core has both a web app and an android app to make your website easy and always available. It offers you all the designs in collaboration with some smart people. Your projects will look great everywhere you go. Use new components that come included!
+      </p>
+      <p>
+        Core is an advanced theme solution for desktop, tablet, and mobile devices. It works in all browsers and your clients can pay instantly, to improve their experience at all times anywhere they go.
+      </p>
+      <div class="icons">
+        <i class="fa fa-desktop fa-3x"></i>
+        <i class="fa fa-tablet fa-3x"></i>
+        <i class="fa fa-mobile-phone fa-3x"></i>
+      </div>
+    </section>
+  </div>
+
+  <div class="spacial-client-quotes">
+    <section class="container">
+      <h2>
+        Used by the world's largest companies
+      </h2>
+      <div>
+        <div class="client-quotes">
+          <div class="client-quote active">
+            <div class="quote">
+              "Focus on creating and growing your projects and websites, and we‘ll take care of spinning up new designs for your users and making sure they’re great. Work with all types of customers just right out of the box!"
+            </div>
+            <div class="author">
+              <span class="name">Jim Warner</span>
+              <span class="role">CTO, Dribbble</span>
+            </div>
+          </div>
+          <div class="client-quote">
+            <div class="quote">
+              "Bring all your ideas together and make customizations faster and better than ever. It works in all browsers and devices plus your clients can pay instantly, to improve their experience at all times anywhere they go."
+            </div>
+            <div class="author">
+              <span class="name">Katia Stevens</span>
+              <span class="role">Marketing Director, Twitter</span>
+            </div>
+          </div>
+          <div class="client-quote">
+            <div class="quote">
+              "Focus on creating and growing your projects and websites, and we‘ll take care of spinning up new designs for your users and making sure they’re great. Work with all types of customers just right out of the box!"
+            </div>
+            <div class="author">
+              <span class="name">John Roberts</span>
+              <span class="role">Product Designer, Github</span>
+            </div>
+          </div>
+          <div class="client-quote">
+            <div class="quote">
+              "Bring all your ideas together and make customizations faster and better than ever. It works in all browsers and devices plus your clients can pay instantly, to improve their experience at all times anywhere they go."
+            </div>
+            <div class="author">
+              <span class="name">Ana Jones</span>
+              <span class="role">Designer, Kickstarter</span>
+            </div>
+          </div>
+          <div class="client-quote">
+            <div class="quote">
+              "Easily customize any part of your project for best results and quick great prototyping. Showcase your work on any computer or mobile device and all screen sizes. Bring all your ideas together and make customizations faster and better than ever."
+            </div>
+            <div class="author">
+              <span class="name">Robert Williams</span>
+              <span class="role">Marketing, Atlassian</span>
+            </div>
+          </div>
+        </div>
+        <div class="client-pictures">
+          <div class="client-picture active">
+            <img src="~theme-corporate/images/uifaces/16.jpg">
+          </div>
+          <div class="client-picture">
+            <img src="~theme-corporate/images/uifaces/13.jpg">
+          </div>
+          <div class="client-picture">
+            <img src="~theme-corporate/images/uifaces/17.jpg">
+          </div>
+          <div class="client-picture">
+            <img src="~theme-corporate/images/uifaces/11.jpg">
+          </div>
+          <div class="client-picture">
+            <img src="~theme-corporate/images/uifaces/14.jpg">
+          </div>
+        </div>
+      </div>
+    </section>
+  </div>
+
+  <div class="call-to-action-signup">
+    <div class="container">
+      <h3>
+        Try it free for 30 days
+      </h3>
+      <form class="form-inline justify-content-center">
+        <div class="form-group">
+          <input type="text" class="form-control" placeholder="Your name" />
+        </div>
+        <div class="form-group">
+          <input type="email" class="form-control" placeholder="Your email address" />
+        </div>
+        <div class="form-group">
+          <input type="password" class="form-control" placeholder="Your password" />
+        </div>
+
+        <button class="btn-shadow btn-shadow-primary" type="submit">
+          Create account
+        </button>
+
+        <div class="agree w-100">By signing up you agree to our <a href="#">Terms</a>.</div>
+      </form>
+>>>>>>> 76b751f... Fix background-images on home page
     </div>
 
     <div class="client-pictures">
