+++ src/views/project.html
@@ -1,0 +1,67 @@
+<div>
+  <section id="project-hero">
+    <div class="inner-bg">
+      <img src="~theme-corporate/images/unsplash/photo-1474843148229-3163319fcc4c.jpg" onload="document.getElementById('project-hero').className += ' show'" />
+    </div>
+    <section>
+      <h1>
+        Online Portfolio Project for Client
+      </h1>
+    </section>
+  </section>
+
+  <section id="project-intro">
+    <div class="container">
+      <p>
+        Focus on creating and growing your projects and websites, and we‘ll take care of spinning up new designs for your users and making sure they’re great. If you’d prefer to customize your projects directly, Spacial can give you unmatched control for customization and work with all types of customers.
+      </p>
+      <p>
+        Spacial has both a web app and an android app to make your website easy and always available. It offers you all the designs in collaboration with some smart people.
+      </p>
+      <p>
+        Spacial is the perfect solution for web developers and designers. You can get an awesome website for your best projects and clients.
+      </p>
+      <div class="devices">
+        <img src="~theme-corporate/images/project/devices.png" class="img-fluid mx-auto d-block">
+      </div>
+    </div>
+  </section>
+
+  <section id="project-showcase">
+    <div class="container">
+      <h3>Online Payments</h3>
+      <p>
+        Spacial has both a web app and an android app to make your website easy and always available. It offers you all the designs in collaboration with some smart people. Your projects will look great everywhere you go. Use new components that come included.  It works in all browsers and your clients can pay instantly, to improve their experience at all times anywhere they go.
+      </p>
+      <p>
+        Focus on creating and growing your projects and websites, and we‘ll take care of spinning up new designs for your users and making sure they’re great. If you’d prefer to customize your projects directly, Spacial can give you unmatched control for customization and work with all types of customers.
+      </p>
+      <div class="browser">
+        <img src="~theme-corporate/images/project/browser.png" class="img-fluid d-block mx-auto">
+      </div>
+      <p>
+        Focus on creating and growing your projects and websites, and we‘ll take care of spinning up new designs for your users and making sure they’re great. If you’d prefer to customize your projects directly, Spacial can give you unmatched control for customization and work with all types of customers.
+      </p>
+      <p>
+        Spacial is the perfect solution for web developers and designers. You can get an awesome website for your best projects and clients.
+      </p>
+      <div class="browser2">
+        <img src="~theme-corporate/images/project/browser2.png" class="img-fluid d-block mx-auto">
+      </div>
+    </div>
+  </section>
+
+  <section id="project-bottom">
+    <div class="link-to-site">
+      <a href="#">www.spacialproject.com</a>
+    </div>
+    <div class="paging clearfix">
+      <div class="previous float-left">
+        <a href="#">« Previous project</a>
+      </div>
+      <div class="next float-right">
+        <a href="#">Next project »</a>
+      </div>
+    </div>
+  </section>
+</div>
